pyproject.toml
[project]
name = "readmethis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "einops>=0.8.2",
    "huggingface-hub[hf-xet]>=0.36.0",
    "mlx-audio>=0.2.9",
    "mlx-lm>=0.29.1",
    "scipy>=1.17.0",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]
